requirements.txt
streamlit
pandas
Pillow
pymilvus
llama-index
llama-index-vector-stores-milvus
pymilvus
llama-index-llms-openai
llama-index-readers-file
